Adding an Admin route for debug
This path will allow us to test sending the messages from the client to
the graphql setup. In the start screen we'll be listening for keypresses
and triggering these messages with logic and timing. This allows us to
shortcut that and ensure that the server is reacting correctly to the
messages.